+++ pkg/generator/gen_deps_test.go
@@ -29,13 +29,4 @@ func TestGenGopkg(t *testing.T) {
 	if gopkgTomlTmpl != buf.String() {
 		t.Errorf(errorMessage, gopkgTomlTmpl, buf.String())
 	}
-
-	buf = &bytes.Buffer{}
-	if err := renderGopkgLockFile(buf); err != nil {
-		t.Error(err)
-		return
-	}
-	if gopkgLockTmpl != buf.String() {
-		t.Errorf(errorMessage, gopkgLockTmpl, buf.String())
-	}
 }