+++ pkg/generator/test_constants.go
@@ -10,5 +10,5 @@ const (
 	appVersion     = "v1alpha1"
 	appGroupName   = "app.example.com"
 	appProjectName = "app-operator"
-	errorMessage   = "Want:\n%vGot:\n%v"
+	errorMessage   = "Want:\n%v\nGot:\n%v"
 )